requirements.txt
streamlit
pandas
duckdb
faker
openai
langchain
python-dotenv
